requirements.txt
pypresence
yandex-music
psutil
pystray
pillow
winsdk